chore: move auditor package under client directory
