+++ pkg/client/auditor/rootservice.go
